feat: CustomLinkedList

+++ src/ru/practicum/task_tracker/manager/HistoryManager.java
@@ -6,6 +6,6 @@ import java.util.List;
 
 public interface HistoryManager {
     void add(Task task);
-    void remove(int id);
+    void remove(long id);
     List<Task> getHistory();
 }
